0x05 NC atoi try 3

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,8 +10,8 @@ int _atoi(char *s)
 {
 	int i = 0, value = 1;
 	int sgn = sign(s);
-	long sum = 0;
-	long prev_sum = 0;
+	unsigned int sum = 0;
+	unsigned int prev_sum = 0;
 
 	while (*(s + i) != '\0')
 	{
